Port crm_phone, hr_phone, hr_recruitment_phone, event_phone to v10
Remove crm_claim_phone, because there is no crm_claim module in v10
Start to port asterisk_click2dial (JS code still needs to be ported)
