+++ crm_phone/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 
-from . import crm_phone
+from . import models
 from . import wizard
