+++ crm_phone/view/crm_phonecall.xml
@@ -5,7 +5,6 @@
 -->
 
 <odoo>
-<data>
 
 
 <record id="crm_phonecall_form" model="ir.ui.view">
@@ -33,8 +32,8 @@
                         <field name="date"/>
                         <field name="user_id"/>
                         <field name="partner_id"/>
-                        <field name="partner_phone" widget="phone"/>
-                        <field name="partner_mobile" widget="phone"/>
+                        <field name="partner_phone"/>
+                        <field name="partner_mobile"/>
                         <field name="opportunity_id"/>
                     </group>
                     <group name="right">
@@ -127,7 +126,7 @@
 </record>
 
 <menuitem id="crm_phonecall_menu" name="Phone Calls"
-      parent="base.menu_base_partner" sequence="7" groups="base.group_sale_salesman"/>
+      parent="sales_team.menu_base_partner" sequence="7" groups="sales_team.group_sale_salesman"/>
 
 <record id="crm_phonecall_action" model="ir.actions.act_window">
     <field name="name">Phone Calls</field>
@@ -156,5 +155,4 @@
     parent="crm_phonecall_menu" sequence="20"/>
 
 
-</data>
 </odoo>
