+++ crm_phone/view/res_partner.xml
@@ -5,14 +5,13 @@
 -->
 
 <odoo>
-<data>
 
 
 <record id="view_partner_form" model="ir.ui.view">
     <field name="name">phonecall.res.partner.form</field>
     <field name="model">res.partner</field>
     <field name="inherit_id" ref="base.view_partner_form"/>
-    <field name="groups_id" eval="[(4, ref('base.group_sale_salesman')), (4, ref('base.group_partner_manager'))]"/>
+    <field name="groups_id" eval="[(4, ref('sales_team.group_sale_salesman')), (4, ref('base.group_partner_manager'))]"/>
     <field name="arch" type="xml">
         <div name="button_box" position="inside">
             <button class="oe_inline oe_stat_button" type="action"
@@ -26,5 +25,4 @@
 </record>
 
 
-</data>
 </odoo>
