+++ crm_phone/wizard/create_crm_phonecall.py
@@ -2,7 +2,7 @@
 # © 2012-2016 Akretion (Alexis de Lattre <alexis.delattre@akretion.com>)
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from openerp import models, api, _
+from odoo import models, api, _
 import phonenumbers
