+++ crm_phone/wizard/create_crm_phonecall_view.xml
@@ -1,12 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <!--
-  Copyright (C) 2012-2013 Akretion (http://www.akretion.com)
-  @author: Alexis de Lattre <alexis.delattre@akretion.com>
-  The licence is in the file __openerp__.py
+  © 2012-2016 Akretion (Alexis de Lattre <alexis.delattre@akretion.com>)
+  License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 -->
 
-<openerp>
-<data>
+<odoo>
 
 <!-- Add button on the "Get partner from incoming phone call view" -->
 <record id="create_crm_phonecall_view" model="ir.ui.view">
@@ -32,5 +30,4 @@
 </record>
 
 
-</data>
-</openerp>
+</odoo>
